Refactor: extract FooterTail from ParquetMetadataReader
